Rewrote by_loc query code to be MUCH simpler
This is part of a database redesign we're involved in. This will
remove the need to do most (all?) of the aggregate pipeline queries
in code.

Also rewrote parseLngLat in terms of Data.Text.split, it's much
simpler now.

Also added some INFO level logging of the by_loc query string
parameters.