Added WAI middleware request logging as a test
This works, and is integrated with the older hslogger logging by
way of having them both log to stdout instead of a file. But I'm
not happy with it. Pretty sure the hslogger logging will fall
apart/behave badly when this thing scales up to multiple cores.

So, committing these changes to record what they looked like at this
time. Moving on to try to use fast-logging directly to a file path
instead. The idea is to move away from hslogger altogether and log my
own custom messages directly to the underlying fast-logger perhaps.