Upgraded to modern 2.x version of stack
This required significant changes to the stack.yaml file and the
addition of stack.yaml.lock to source control.

This change also includes changing the source code layout and the
addition of a library of common code (currently just the APIKey code) to
reduce redundant compilation.

The project was also switched over from explicity cabal file
configuration to hpack and the hs-rest.cabal file was converted into the
package.yaml file.

Updated README to reflect the new build instructions

Finally, the resources were moved into the pack/ directory which allows
to now build for deployment with version 2.5+ of hsinstall.